Fixing occasional trouble loading anytabs
Since Jcycle was running but no code was being loaded, it would
occasionally load the content after. Though it would often get caught up
and not load the content after jcycle WideSlideshow, this included
anytabs. This is because jcycle was set to 1 instead of 0, now set to 0
and jquery for slideshow is commented out.